# quizano/api/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Quizano Simple App API
  version: 1.0.0
  description: Node.js API for a simplified Quizano app with student and admin flows.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register student account
      description: Register a new student user and return access token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              fullName: Nguyen Van C
              username: nguyenvanc
              email: nguyenvanc@gmail.com
              password: '123456'
      responses:
        '201':
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login
      description: Login by username and password. Optionally restrict role by passing role in body.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: admin
              password: '123'
              role: admin
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current authenticated user
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/admin/results:
    get:
      tags: [Admin]
      summary: Query result records with pagination
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: examId
          schema:
            type: string
        - in: query
          name: studentId
          schema:
            type: string
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Paginated result list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminResultsResponse'
  /api/v1/admin/stats/summary:
    get:
      tags: [Admin]
      summary: Score summary and distribution
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: examId
          schema:
            type: string
      responses:
        '200':
          description: Summary metrics
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/StatsSummary'
  /api/v1/admin/students/{studentId}/journey:
    get:
      tags: [Admin]
      summary: Student exam journey
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: studentId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Student journey response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentJourneyResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/admin/students/{studentId}/results/{resultId}:
    get:
      tags: [Admin]
      summary: Admin view detailed result of one student
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: studentId
          required: true
          schema:
            type: string
        - in: path
          name: resultId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detailed result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/AdminDetailedResult'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            ok: false
            error:
              message: Validation failed
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            ok: false
            error:
              message: Invalid or expired access token
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            ok: false
            error:
              message: Resource not found
    Conflict:
      description: Conflict state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            ok: false
            error:
              message: Username already exists
  schemas:
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
            details:
              nullable: true
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, admin]
    AuthSuccessResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        data:
          type: object
          properties:
            token:
              type: string
            user:
              $ref: '#/components/schemas/User'
    RegisterRequest:
      type: object
      required: [fullName, username, email, password]
      properties:
        fullName:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 3
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [student, admin]
...
    Exam:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [free, scheduled]
        duration:
          type: integer
        status:
          type: string
          enum: [active, inactive]
        startTime:
          type: string
          format: date-time
          nullable: true
        endTime:
          type: string
          format: date-time
          nullable: true
...
    Option:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
...
    AnswerInput:
      type: object
      required: [questionId]
      properties:
        questionId:
          type: string
        selectedOptionId:
          type: string
          nullable: true
...
    Result:
      type: object
      properties:
        id:
          type: string
        studentId:
          type: string
        examId:
          type: string
        startTime:
          type: string
          format: date-time
        submitTime:
          type: string
          format: date-time
        answers:
          type: array
          items:
            $ref: '#/components/schemas/AnswerInput'
        correctCount:
          type: integer
        totalQuestions:
          type: integer
        score:
          type: number
        status:
          type: string
          example: completed
...
    AdminResultItem:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            examTitle:
              type: string
              nullable: true
            studentName:
              type: string
              nullable: true
            username:
              type: string
              nullable: true
            pass:
              type: boolean
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        totalItems:
          type: integer
        totalPages:
          type: integer
    AdminResultsResponse:
      type: object
      properties:
        ok:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/AdminResultItem'
            pagination:
              $ref: '#/components/schemas/Pagination'
    StatsDistribution:
      type: object
      properties:
        below3:
          type: integer
        from3to5:
          type: integer
        from5to7:
          type: integer
        from7to9:
          type: integer
        aboveOrEqual9:
          type: integer
    StatsSummary:
      type: object
      properties:
        totalAttempts:
          type: integer
        avgScore:
          type: number
        passRate:
          type: number
        distribution:
          $ref: '#/components/schemas/StatsDistribution'
    StudentJourneyResponse:
      type: object
      properties:
        ok:
          type: boolean
        data:
          type: object
          properties:
            student:
              $ref: '#/components/schemas/User'
            results:
              type: array
              items:
                $ref: '#/components/schemas/AdminResultItem'
    AdminDetailedResultQuestion:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        options:
          type: array
          items:
            $ref: '#/components/schemas/Option'
        explanation:
          type: string
        correctOptionId:
          type: string
        selectedOptionId:
          type: string
          nullable: true
        isCorrect:
          type: boolean
    AdminDetailedResult:
      allOf:
        - $ref: '#/components/schemas/Result'
        - type: object
          properties:
            exam:
              $ref: '#/components/schemas/Exam'
            questions:
              type: array
              items:
                $ref: '#/components/schemas/AdminDetailedResultQuestion'